+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 FEATURES:
 
+* Add a new optional field `playbook` to resource/sumologic_monitor.
 * Add a new optional field `evaluation_delay` to resource/sumologic_monitor.
 
 ## 2.9.9 (August 12, 2021)